refactor: combine and reuse some components
