+++ src/atoms/FigCaption.js
@@ -1,9 +1,32 @@
-import styled from 'react-emotion';
+import styled, { css } from 'react-emotion';
+
+import { getColor, getVariable } from '../utils';
 
 const FigCaption = styled.figcaption`
-	font-size: ${props => props.theme.variables.uiRegularSize};
-	line-height: ${props => props.theme.variables.uiRegularLineHeight};
-	margin-bottom: ${props => props.theme.variables.verticalBase};
+	${(props) => {
+		const { border } = props;
+
+		const grayTintLighter = getColor('grayTintLighter')(props);
+		const uiSmallSize = getVariable('uiSmallSize')(props);
+		const uiSmallLineHeight = getVariable('uiSmallLineHeight')(props);
+		const uiWeight = getVariable('uiWeight')(props);
+		const verticalBase = getVariable('verticalBase')(props);
+
+		return css`
+			display: block;
+			margin: calc(1/2 * ${verticalBase}) 0;
+			padding: 0;
+			${border && css`
+				border-bottom: .1rem solid ${grayTintLighter};
+				padding-bottom: calc(1/2 * ${verticalBase});
+			`}
+			max-width: 100%;
+			font-size: ${uiSmallSize};
+			line-height: ${uiSmallLineHeight};
+			font-weight: ${uiWeight};
+			z-index: 1;
+		`;
+	}}
 `;
 
 export { FigCaption };
